Fix maintainVisibleContentPosition for reverse flexDirections (#47159)
Summary:

The existing `maintainVisibleContentPosition` implementation assumes top-to-bottom layout on iOS. It selected the fixed position item as the first child in the scroll content container whose offset was larger than the current offset.

However, for lists that render children in reverse order, e.g., with `flexDirection: 'column-reverse'`, the first child will almost always be chosen as the fixed position item. In these reversed list cases, we need to choose the first item with an offset less than the bottom of the viewport.

## Changelog

[iOS][Fixed] Fixes `maintainVisibleContentPosition` for reverse ordered lists

Differential Revision: D64575696