Close a view leak due to lossy onAnimationEnd callback
Android's onAnimationEnd callback is lossy and ocasionally just does not
fire.  However the LayoutAnimationController maintains a sparse array of
animations (with Strong View refs) that is only cleaned when the
onAnimationEnd callback is invoked.  This results in a leak of Android
View objects over time.

To avoid this, the Strong View refs are migrated to WeakReference's and
the associated sparse array is cleaned of any invalid layout animations
in response to the reset() call.

This closes two leaks:
1. Unbound growth in LayoutAnimationController::mLayoutHandlers
2. Pinning View objects into memory as the sole remaining GC root