Change packages resolution for buildTypes script
Summary:
Changelog:
[internal]


The simpleResolve will check for deep imports like `react-native/assets-registry/registry`. The "main" property also has to be added to the assets's package.json to indicate the entry point (which is used in the simpleResolve algorithm). This will generate types_generated folder for matched packages outside of react-native.

Differential Revision: D70974073