[ios][prebuild] fixed inclusion of resources in swift package
We had some issues with the Swift package build step where we saw an error message when we included resources and couldn't find out why this was happening.

After systematically going through the generated swift package file and looking for a reason I found a mistake.

When we generate the Package.swift file we pass all compilerFlags from the configuration of the target to both cpp/c flags - which in the case of the folly target ends up being passed to the dependency scanner which isn't too happy about this c++ flag.

The solution is to break out the c++ stdlib version as a separate setting in the configuration and pass it only to the `cxxSettings` section in our final Package.swift file.

This commit fixes this by:
- Adding a new field in the dependency settings struct called `cppVersion`
- Updated configuration with correct resources and cppVersion for relevant targets
- Now emits the correct cppVersion to the cxxSettings section
- Fixed issue with the copy bundles step that didn't copy the directory in some cases.