Add castInterface API in JSI (#49604)
Summary:

Add JSI castInterface API. This provides a structured way to provide
optional APIs that may or may not be supported in all engines.

Every interface will have its own set of APIs. For instance, JSI Runtime
will provide a set of APIs that all JSI Runtimes should support. Hermes
will provide an interface (IHermesRuntime) containing Hermes-specific
APIs. In the future, an interface may be created to expose optional
APIs.

A runtime will inherit the interfaces of which the APIs are supported by
the runtime. Users who want to access a specific API may call
castInterface with the UUID of interface containing the API.

Changelog: [Internal]

Differential Revision: D68964360