Add SyncAccessibilityInfo TM, SyncAccessibilityManager TM and useIsScreenReaderEnabled hook (#50138)
Summary:

This diff implements two new Turbo Modules to enabled the synchronous query of Accessibility information from both Android and iOS.

As a first step this also implements a new hook `useIsScreenReaderEnabled` this would the the first hook in the series of accessibility hooks to enabled synchronous fetching of accessibility information from native.

These hooks will allow you to subscribe to the accessibility information triggering automatic re-renders with the new information if subscribed.

For example: 
```
const isScreenReaderEnabled = useIsScreenReaderEnabled()
```
will subscribe to the state of the screen reader and update the value of it accordingly if the state changes at any point.

Also added a feature flag `enableAccessibilityHooks` to gate this feature during development

Differential Revision: D70917330