Remove `__REACT_DEVTOOLS_GLOBAL_HOOK__ ` from react-native libdef
Summary:
This is an implementation detail of React and better not make it leak everywhere. It's not used quite often even in react-native codebase, so it's better to just suppress the error on each call site. Currently it's typed as any, so there is not much type safety lost anyways.

Changelog: [Internal]

Differential Revision: D71687426