chore: add jest-config to monorepo dependencies (#50280)
Summary:
The React Native monorepo depends on `jest-config` in `jest.config.js` but does not specify it as a dependency. This means we got it as a phantom / transitive dependency. In React Native macOS, I am testing using Yarn 4 with pnpm layout to protect against such dependencies (See https://github.com/microsoft/react-native-macos/pull/2366). The simplest way to fix this is to just declare it as a dependency.

## Changelog:

[INTERNAL] [FIXED] - Add jest-config as a dependency

Pull Request resolved: https://github.com/facebook/react-native/pull/50280

Test Plan: This change should be a no-op in React Native, we already had the package in our lock.

Reviewed By: yungsters, huntie

Differential Revision: D71891841

Pulled By: robhogan

fbshipit-source-id: 5b4720e9e09e25b2f1eebb80675bfe4f3a48e2dc