Remove ANDROID ifdefs from TextShadowNode
Summary:
Noticed this logic from early Fabric, and it seems suppper funky in the context of how the rest of the systems work today. `Text` doesn't form a view (compared to `Paragraph`), and Differentiator will treat non-layoutable, or shadownodes with empty layout metrics as virtual, so I can't think of why this would do anything. Let's clean it up.

Changelog: [Internal]

Differential Revision: D71941993