Use smart pointers and add missing virtual destructor
Summary:
# Changelog: [Internal]

We have different types of call frames, right now we defined 4 of them. They all descend from `ProfileSampleCallStackFrame`, which has `kind_` field that can be used for determining the type of the call frame.

We are using polymorphism here, so that later we can cast base type to sub type, based on the `kind`. This happens on React Native side in `HermesRuntimeSamplingProfileSerializer.cpp`, where we transform Hermes' payload into more CDP-like structures in React Native domain.

Because of polymorphism, not having a virtual destructor might have been problematic, since we are storing the vector of pointers of Base type. My current understanding is that this may cause undefined behaviour and when vector of base pointers is cleared, the resources allocated by the sub type would not be cleared.

Instead of storing raw pointers, we will use smart unique pointer. `ProfileGenerator.cpp` is a simple serializer, so it will just transfer the ownership of call frames once it returns the JavaScript samples payload. Because of that, some structs are made only movable and not copyable.

Differential Revision: D71993763