Migrate to Kotlin - DevSupportManagerBase.kt (#50559)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/50559

This diff migrates the following file to Kotlin - DevSupportManagerBase.kt
as part of our ongoing effort of migrating the codebase to Kotlin

This file was 1000LOC+ so its migration is quite involved.
I had to look into nullability of various parameters. Most notably the `Array<StackFrame?>` -> `Array<StackFrame>`.

This also is the last file left in the devsupport package that needs to be migrated to Kotlin.

Changelog:
[Android] [Breaking] - DevSupportManagerBase is now converted to Kotlin. If you're subclassing this class, you will have to adjust some of the parameters as types have changed during the migration.

Reviewed By: javache, mdvacca

Differential Revision: D72637098

fbshipit-source-id: b87d5f17e30d69fe9c09ec9b49f96818df34ea3b